+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/AGE_1.01.iml" filepath="$PROJECT_DIR$/AGE_1.01.iml" />
+      <module fileurl="file://$PROJECT_DIR$/AgeClub.iml" filepath="$PROJECT_DIR$/AgeClub.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
