+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="age.club.administrator.age_101">
+    package="age.AgeClub.administrator.age_102">
 
     <!-- To auto-complete the email text field in the login form with the user's emails -->
     <uses-permission android:name="android.permission.GET_ACCOUNTS" />
@@ -27,7 +27,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name="age.club.administrator.age_101.LoginActivity"
+            android:name="age.AgeClub.administrator.age_102.LoginActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -36,7 +36,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="age.club.administrator.age_101.MainActivity"
+            android:name="age.AgeClub.administrator.age_102.MainActivity"
             android:label="@string/title_activity_main"
             android:theme="@style/AppTheme.NoActionBar"
             android:configChanges="keyboardHidden|orientation|screenSize"/>
@@ -53,7 +53,7 @@
             android:value="@string/google_maps_key" />
 
         <activity
-            android:name="age.club.administrator.age_101.MapActivity"
+            android:name="age.AgeClub.administrator.age_102.MapActivity"
             android:label="@string/title_activity_map"/>
     </application>
