--- app/src/main/java/age/club/administrator/age_101/QrTransaction.java
@@ -1,33 +1,0 @@
-package age.club.administrator.age_101;
-
-
-public class QrTransaction {
-
-    private String customerID;
-    private String companyID;
-    private String time;
-
-    public String getCustomerID() {
-        return customerID;
-    }
-
-    public void setCustomerID(String customerID) {
-        this.customerID = customerID;
-    }
-
-    public String getCompanyID() {
-        return companyID;
-    }
-
-    public void setCompanyID(String companyID) {
-        this.companyID = companyID;
-    }
-
-    public String getTime() {
-        return time;
-    }
-
-    public void setTime(String time) {
-        this.time = time;
-    }
-}
