+++ app/src/main/res/layout/activity_login.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="age.club.administrator.age_101.LoginActivity">
+    tools:context="age.AgeClub.administrator.age_102.LoginActivity">
 
     <!-- Login progress -->
     <ProgressBar
