+++ app/src/main/res/layout/app_bar_main.xml
@@ -5,7 +5,7 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="age.club.administrator.age_101.MainActivity">
+    tools:context="age.AgeClub.administrator.age_102.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
