+++ app/src/main/res/layout/content_discounts.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="age.club.administrator.age_101.MainActivity"
+    tools:context="age.AgeClub.administrator.age_102.MainActivity"
     tools:showIn="@layout/app_bar_main">
 
     <Spinner
