+++ app/src/main/res/layout/content_map.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="age.club.administrator.age_101.MainActivity">
+    tools:context="age.AgeClub.administrator.age_102.MainActivity">
     tools:showIn="@layout/app_bar_main">
 
     <TextView
