+++ app/src/main/res/layout/content_personal_info.xml
@@ -12,7 +12,7 @@
         android:layout_height="match_parent"
         android:layout_margin="25dp"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        tools:context="age.club.administrator.age_101.MainActivity">
+        tools:context="age.AgeClub.administrator.age_102.MainActivity">
 
         <TextView
             android:id="@+id/textView1"
