+++ app/src/release/res/values/google_maps_api.xml
@@ -17,6 +17,6 @@
     string in this file.
     -->
     <string name="google_maps_key" templateMergeStrategy="preserve" translatable="false">
-        AIzaSyBoMKYUMeOkRjwrYb7ji5Pa05gikEv-3Ec
+        AIzaSyCSRlauC3RiA633EIdxf7ncc5xoDUzCrgQ
     </string>
 </resources>
